Test admin view instead of api view
